move app files to addon folder
