+++ addon/instance-initializers/theme-init.js
@@ -1,0 +1,10 @@
+export function initialize(applicationInstance) {
+  applicationInstance
+    .lookup('service:theme-changer')
+    ._generateStyleTag();
+}
+
+export default {
+  name: 'ember-theme-changerr-init',
+  initialize
+};
